match/requirements.txt
redis==2.10.6
kafka-python==1.4.3